updated framework to JUCE v3.2.0

+++ Source/common/parameter_juggler/plugin_parameter_continuous.cpp
@@ -345,7 +345,7 @@ const String PluginParameterContinuous::getTextFromFloat(float newValue)
     else
     {
         // round real parameter value
-        newRealValue = MZ_Juce_Common::round_mz(newRealValue);
+        newRealValue = (float) MZ_Juce_Common::round_mz(newRealValue);
 
         // format parameter value
         textValueNew = String(newRealValue);
